# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stone World API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface AdminSessionResponse {
  authenticated: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  sortOrder: number;
  productCount: number;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  description?: string | null;
  sortOrder?: number;
}

export type ProductOrigin = (typeof ProductOrigin)[keyof typeof ProductOrigin];

export const ProductOrigin = {
  imported: "imported",
  national: "national",
} as const;

export interface Product {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  categoryId?: number | null;
  categoryName?: string | null;
  price?: number | null;
  priceUnit?: string | null;
  origin: ProductOrigin;
  available: boolean;
  featured: boolean;
  imageUrl?: string | null;
  images: string[];
  tags: string[];
  deliveryAvailable: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateProductBodyOrigin =
  (typeof CreateProductBodyOrigin)[keyof typeof CreateProductBodyOrigin];

export const CreateProductBodyOrigin = {
  imported: "imported",
  national: "national",
} as const;

export interface CreateProductBody {
  name: string;
  description?: string | null;
  categoryId?: number | null;
  price?: number | null;
  priceUnit?: string | null;
  origin: CreateProductBodyOrigin;
  available?: boolean;
  featured?: boolean;
  imageUrl?: string | null;
  images?: string[];
  tags?: string[];
  deliveryAvailable?: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  limit: number;
  offset: number;
}

export interface BulkDeleteBody {
  ids: number[];
}

export interface BulkToggleBody {
  ids: number[];
  available: boolean;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt?: string | null;
  content: string;
  coverImageUrl?: string | null;
  published: boolean;
  aiGenerated: boolean;
  tags: string[];
  readTimeMinutes: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBlogPostBody {
  title: string;
  excerpt?: string | null;
  content: string;
  coverImageUrl?: string | null;
  published?: boolean;
  tags?: string[];
}

export interface GenerateBlogPostBody {
  topic: string;
  keywords?: string[];
  publish?: boolean;
}

export interface BlogListResponse {
  posts: BlogPost[];
  total: number;
  limit: number;
  offset: number;
}

export type EnquiryAudience =
  (typeof EnquiryAudience)[keyof typeof EnquiryAudience];

export const EnquiryAudience = {
  homeowner: "homeowner",
  contractor: "contractor",
  architect: "architect",
  developer: "developer",
} as const;

export interface Enquiry {
  id: number;
  name: string;
  email?: string | null;
  phone: string;
  audience: EnquiryAudience;
  projectType?: string | null;
  budget?: string | null;
  message: string;
  productInterest?: string | null;
  location?: string | null;
  isRead: boolean;
  createdAt: string;
}

export type CreateEnquiryBodyAudience =
  (typeof CreateEnquiryBodyAudience)[keyof typeof CreateEnquiryBodyAudience];

export const CreateEnquiryBodyAudience = {
  homeowner: "homeowner",
  contractor: "contractor",
  architect: "architect",
  developer: "developer",
} as const;

export interface CreateEnquiryBody {
  name: string;
  email?: string | null;
  phone: string;
  audience: CreateEnquiryBodyAudience;
  projectType?: string | null;
  budget?: string | null;
  message: string;
  productInterest?: string | null;
  location?: string | null;
}

export interface MarkReadBody {
  isRead: boolean;
}

export interface EnquiryListResponse {
  enquiries: Enquiry[];
  total: number;
  unread: number;
  limit: number;
  offset: number;
}

export interface UploadMediaBody {
  filename: string;
  dataUrl: string;
  mimeType: string;
}

export interface MediaItem {
  id: number;
  filename: string;
  url: string;
  mimeType: string;
  size: number;
  createdAt: string;
}

export interface SiteSettings {
  id: number;
  companyName: string;
  address?: string | null;
  phone?: string | null;
  email?: string | null;
  whatsapp?: string | null;
  instagramUrl?: string | null;
  facebookUrl?: string | null;
  adminPasswordHash: string;
  updatedAt: string;
}

export interface UpdateSettingsBody {
  companyName?: string;
  address?: string | null;
  phone?: string | null;
  email?: string | null;
  whatsapp?: string | null;
  instagramUrl?: string | null;
  facebookUrl?: string | null;
  currentPassword?: string | null;
  newPassword?: string | null;
}

export interface DashboardStats {
  totalProducts: number;
  availableProducts: number;
  totalEnquiries: number;
  unreadEnquiries: number;
  totalBlogPosts: number;
  publishedBlogPosts: number;
  totalCategories: number;
  totalMedia: number;
  recentEnquiries: Enquiry[];
}

export interface CategoryStat {
  categoryId?: number | null;
  categoryName: string;
  count: number;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  available?: boolean;
  origin?: ListProductsOrigin;
  limit?: number;
  offset?: number;
};

export type ListProductsOrigin =
  (typeof ListProductsOrigin)[keyof typeof ListProductsOrigin];

export const ListProductsOrigin = {
  imported: "imported",
  national: "national",
  all: "all",
} as const;

export type ListBlogPostsParams = {
  published?: boolean;
  limit?: number;
  offset?: number;
};

export type ListEnquiriesParams = {
  audience?: ListEnquiriesAudience;
  read?: boolean;
  limit?: number;
  offset?: number;
};

export type ListEnquiriesAudience =
  (typeof ListEnquiriesAudience)[keyof typeof ListEnquiriesAudience];

export const ListEnquiriesAudience = {
  homeowner: "homeowner",
  contractor: "contractor",
  architect: "architect",
  developer: "developer",
  all: "all",
} as const;

export type ExportEnquiriesParams = {
  audience?: string;
};