# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stone World API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Products ────────────────────────────────────────────────────────────────
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog articles
      parameters:
        - name: published
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post manually (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/generate:
    post:
      operationId: generateBlogPost
      tags: [blog]
      summary: AI-generate a blog article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBlogPostBody"
      responses:
        "201":
          description: Generated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "200":
          description: Blog post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Enquiries ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Admin ─────────────────────────────────────────────────────────────────
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required:
        - success
        - token
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - sortOrder
        - productCount
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - name

    # ─── Products ──────────────────────────────────────────────────────────────
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        published:
          type: boolean
        aiGenerated:
          type: boolean
        tags:
          type: array
          items:
            type: string
        readTimeMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - content
        - published
        - aiGenerated
        - tags
        - readTimeMinutes
        - createdAt
        - updatedAt
    CreateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        published:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - content
    GenerateBlogPostBody:
      type: object
      properties:
        topic:
          type: string
        keywords:
          type: array
          items:
            type: string
        publish:
          type: boolean
      required:
        - topic
    BlogListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - posts
        - total
        - limit
        - offset

    # ─── Enquiries ─────────────────────────────────────────────────────────────
...